src/Kotas.Utils.Common/Utils/PrimitiveTypeUtils.cs
using System;
using System.Reflection;

namespace Kotas.Utils.Common.Utils
{
    public static class PrimitiveTypeUtils
    {
        public static bool IsPrimitive(Type type)
        {
            var typeInfo = type.GetTypeInfo();
            if (typeInfo.IsGenericType)
            {
                return false;
            }
            return typeInfo.IsPrimitive
                   || typeInfo.IsEnum
                   || type == typeof(string)
                   || type == typeof(decimal)
                   || type == typeof(DateTime);
        }

        public static bool IsPrimitiveWithGeneric(Type type)
        {
            var typeInfo = type.GetTypeInfo();
            if (typeInfo.IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>))
            {
                // nullable type, check if the nested type is primitive
                return IsPrimitiveWithGeneric(typeInfo.GetGenericArguments()[0]);
            }
            return typeInfo.IsPrimitive
                   || typeInfo.IsEnum
                   || type == typeof(string)
                   || type == typeof(decimal)
                   || type == typeof(DateTime);
        }

    }
}
